+++ new-dti-website/components/products/FloatingImages.tsx
